intra_metabolites vector getter

+++ Cell.h
@@ -4,7 +4,7 @@
 //==============================
 //    INCLUDES
 //==============================
-
+#include <vector>
 
 /**
  * @class Cell
@@ -32,6 +32,7 @@ std::vector<float> intra_metabolites();
 //==============================
 
 //==============================
+
 //    OPERATORS
 //==============================
 
@@ -47,7 +48,7 @@ protected:
 //==============================
 //    ATTRIBUTES
 //==============================
-  static const int fitness_min_;
+  static const float fitness_min_;
   static const float Pdeath_;
   static const float Pmut_;
   int fitness_;
